[refactor] (feat) :  domain과 ui 역할 분리 - setup.js / modal.js
