+++ src/view/web/setup.js
